Move usePrepareProductForCart to own file
